Revert the provider indirection logic while keeping the timeout functionality to fix resource notifications and ChefSpec stubs.
