+++ lib/chef/platform/provider_mapping.rb
@@ -64,6 +64,7 @@ class Chef
             },
             :ubuntu   => {
               :default => {
+                :package => Chef::Provider::Package::Apt,
                 :service => Chef::Provider::Service::Debian,
                 :cron => Chef::Provider::Cron,
                 :mdadm => Chef::Provider::Mdadm
@@ -79,6 +80,7 @@ class Chef
             },
             :gcel   => {
               :default => {
+                :package => Chef::Provider::Package::Apt,
                 :service => Chef::Provider::Service::Debian,
                 :cron => Chef::Provider::Cron,
                 :mdadm => Chef::Provider::Mdadm
@@ -86,6 +88,7 @@ class Chef
             },
             :linaro   => {
               :default => {
+                :package => Chef::Provider::Package::Apt,
                 :service => Chef::Provider::Service::Debian,
                 :cron => Chef::Provider::Cron,
                 :mdadm => Chef::Provider::Mdadm
@@ -93,6 +96,7 @@ class Chef
             },
             :raspbian   => {
               :default => {
+                :package => Chef::Provider::Package::Apt,
                 :service => Chef::Provider::Service::Debian,
                 :cron => Chef::Provider::Cron,
                 :mdadm => Chef::Provider::Mdadm
@@ -100,6 +104,7 @@ class Chef
             },
             :linuxmint   => {
               :default => {
+                :package => Chef::Provider::Package::Apt,
                 :service => Chef::Provider::Service::Upstart,
                 :cron => Chef::Provider::Cron,
                 :mdadm => Chef::Provider::Mdadm
@@ -107,6 +112,7 @@ class Chef
             },
             :debian => {
               :default => {
+                :package => Chef::Provider::Package::Apt,
                 :service => Chef::Provider::Service::Debian,
                 :cron => Chef::Provider::Cron,
                 :mdadm => Chef::Provider::Mdadm
